+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Master
+### Added
+- Type hints to the adapter interface
+
 ### Changed
 - Require PHP from 7.2 to 7.0
 - Require PHPUnit from 7 to 6